python/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pytest==7.3.1
requests==2.31.0
PyJWT==2.7.0
